file restructuring and renaming
