+++ resources/assets/components/utilities/PopoverDispatcher/SitewideBanner/SitewideBanner.js
@@ -4,9 +4,9 @@ import PropTypes from 'prop-types';
 import {
   EVENT_CATEGORIES,
   trackAnalyticsEvent,
-} from '../../../helpers/analytics';
-import PrimaryButton from '../Button/PrimaryButton';
-import CloseButton from '../../artifacts/CloseButton/CloseButton';
+} from '../../../../helpers/analytics';
+import PrimaryButton from '../../Button/PrimaryButton';
+import CloseButton from '../../../artifacts/CloseButton/CloseButton';
 
 const SitewideBannerContent = ({
   cta,
