+++ resources/assets/components/utilities/PopoverDispatcher/config.js
@@ -1,11 +1,9 @@
 /**
- * This file may eventually turn into config variables to avoid hardcoding when to suppress.
  *
  * Because the SitewideBanner CTA is hardcoded to redirect to vote.dosomething.org, we want to hide
  * the banner on certain pages based on content, or when authenticated user is already registered or
  * ineligible to vote.
  */
-
 // An asterik can be used to exclude all children of a path, e.g. /us/campaigns/ready-vote/*
 export const excludedPaths = [
   '/us/campaigns/ready-vote',
