+++ resources/views/layouts/master.blade.php
@@ -33,7 +33,7 @@
     @endif
 
     <div id="fb-root"></div>
-    <div id="banner-portal" class="top-0 left-0 w-full block z-1000" role="presentation"></div>
+    <div id="popover-portal" class="top-0 left-0 w-full block z-1000" role="presentation"></div>
     <div id="chrome" class="chrome min-h-screen w-full">
         @yield('content')
     </div>
